pyproject.toml
[tool.poetry]
name = "azx_test"
version = "0.1.0"
description = "Backend Developer Test"
authors = ["Carlos <carlosjcosta1981@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.0.0"
PyJWT = "^2.2.0"
pytest = "^6.2.4"
flask-expects-json = "^1.7.0"
click = "^8.1.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
